+++ internal/backend/backend.go
@@ -1,3 +1,10 @@
 package backend
 
+import "net/http"
+
 // TODO: Create a backend interface
+
+type Backend interface {
+	HandleModelsRequest(w http.ResponseWriter)
+	HandleChatCompletions(w http.ResponseWriter, r *http.Request)
+}
